refactor(login): move the provider login flows into the SDK
The flows sat in apps/pythinker-code, so the VS Code extension could not use
them — it depends on @pythoughts/pythinker-code-sdk alone, and still calls
harness.auth.login(undefined, ...), which offers no provider choice at all.

The SDK is the right home rather than the oauth package: oauth has no
@pythoughts dependencies and must keep none, while the flows need fetchCatalog
and applyCatalogProvider, which live in the SDK. Putting them in oauth would
have inverted that edge into a cycle. Nothing new is added to any package.json.

Two couplings dissolved on inspection rather than being carried across. The
port declared showStatus(message, level) against a TUI theme token, but no flow
ever passes a level, so the parameter and the theme dependency are gone. The
remaining app types (spinner handle, api-key prompt options, platform
selection) are now declared in the SDK and matched structurally. Generic error
formatting moves out of a TUI util into the SDK, re-exported so its existing
importers are untouched.

Browser opening becomes a port member instead of a helper the flows call
directly. It was duplicated byte-for-byte into the SDK during the move, and it
was never renderer-agnostic to begin with: a terminal shells out, while an
editor extension has to use its own host API. Owning it in the port removes the
copy and is what lets a VS Code renderer exist.

Also fixes a flaky test rather than leaving it annotated: the skill telemetry
assertion awaited a skill.activated event and then read telemetry records that
are written independently of it, so a loaded machine lost the race.
