+++ apps/pythinker-code/test/tui/components/dialogs/platform-selector.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
 
-import { CATALOG_PLATFORM_VALUE_PREFIX } from '#/auth/platform-values';
+import { CATALOG_PLATFORM_VALUE_PREFIX } from '@pythoughts/pythinker-code-sdk';
 import { PlatformSelectorComponent } from '#/tui/components/dialogs/platform-selector';
 import { promptPlatformSelection } from '#/tui/commands/prompts';
